// openapi/dres/model/queryResultLog.ts
/**
 * DRES Client API
 * Client API for DRES (Distributed Retrieval Evaluation Server), Version 2.0.3
 *
 * The version of the OpenAPI document: 2.0.3
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RankedAnswer } from './rankedAnswer';
import { QueryEvent } from './queryEvent';


export interface QueryResultLog { 
    timestamp: number;
    sortType: string;
    resultSetAvailability: string;
    results: Array<RankedAnswer>;
    events: Array<QueryEvent>;
}