// openapi/dres/model/successfulSubmissionsStatus.ts
/**
 * DRES API
 * API for DRES (Distributed Retrieval Evaluation Server), Version 1.0
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SuccessfulSubmissionsStatus { 
    submission: SuccessfulSubmissionsStatus.SubmissionEnum;
    description: string;
    status: boolean;
}
export namespace SuccessfulSubmissionsStatus {
    export type SubmissionEnum = 'CORRECT' | 'WRONG' | 'INDETERMINATE' | 'UNDECIDABLE';
    export const SubmissionEnum = {
        CORRECT: 'CORRECT' as SubmissionEnum,
        WRONG: 'WRONG' as SubmissionEnum,
        INDETERMINATE: 'INDETERMINATE' as SubmissionEnum,
        UNDECIDABLE: 'UNDECIDABLE' as SubmissionEnum
    };
}